Update 24_11_25
- fix hog transform core algorithm issue (loss of accuracy) at both version (pc, mcu) but will slow down (x2-x3 time)
- upgrade packed_vector : from 32-bit to unlimited bits per element
- change Tree_node structure: allow 64-bit mode (now can handle large dataset > 1.000.000 samplels)
- Optimize storage for model file.
- fix: node_predictor cannot created with --skip_training argument.
- remove and clean 64-bit tree_node legacy.

+++ src/hog_mcu/hog_transform.h
@@ -239,7 +239,7 @@ private:
     // Optimization: Pre-allocated buffers to avoid repeated allocations
     int16_t* gradient_x_buffer;       // Gradient X components
     int16_t* gradient_y_buffer;       // Gradient Y components
-    uint16_t* magnitude_buffer;       // Gradient magnitudes (integer)
+    float* magnitude_buffer;          // Gradient magnitudes (float)
     uint8_t* angle_bin_buffer;        // Pre-computed angle bins
     float* block_histogram_buffer;    // Reusable block histogram buffer
     float* cell_histogram_buffer;     // Reusable cell histogram buffer
@@ -255,6 +255,7 @@ private:
     
     // Helper methods for gradient computation
     float computeGradientMagnitude(int gx, int gy);
+    // Compute gradient angle in radians (0 .. 2*PI)
     float computeGradientAngle(int gx, int gy);
     
     // Optimized gradient computation using integer arithmetic
